Security context: resolved circular dependency while creating CustomUserDetailsService and AuthenticationManager.
