+++ src/main/java/pl/dreilt/basicspringmvcapp/service/AppUserService.java
@@ -5,7 +5,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.dreilt.basicspringmvcapp.dto.*;
 import pl.dreilt.basicspringmvcapp.entity.AppUser;
@@ -17,7 +16,6 @@ import pl.dreilt.basicspringmvcapp.repository.AppUserRoleRepository;
 
 import java.util.Optional;
 
-@Service
 public class AppUserService {
 
     private static final String USER_ROLE = "USER";
