+++ src/main/java/pl/dreilt/basicspringmvcapp/service/CustomUserDetailsService.java
@@ -1,0 +1,36 @@
+package pl.dreilt.basicspringmvcapp.service;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import pl.dreilt.basicspringmvcapp.dto.AppUserCredentialsDto;
+
+public class CustomUserDetailsService implements UserDetailsService {
+    protected final Log logger = LogFactory.getLog(this.getClass());
+    private final AppUserService appUserService;
+
+    public CustomUserDetailsService(final AppUserService appUserService) {
+        this.appUserService = appUserService;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return appUserService.findAppUserCredentialsByEmail(username)
+                .map(this::createUserDetails)
+                .orElseThrow(() -> new UsernameNotFoundException(String.format("User with email %s not found", username)));
+    }
+
+    private UserDetails createUserDetails(AppUserCredentialsDto appUserCredentialsDto) {
+        return User.builder()
+                .username(appUserCredentialsDto.getEmail())
+                .password(appUserCredentialsDto.getPassword())
+                .disabled(!appUserCredentialsDto.isEnabled())
+                .accountLocked(!appUserCredentialsDto.isAccountNonLocked())
+                .roles(appUserCredentialsDto.getRoles().toArray(String[]::new))
+                .build();
+    }
+
+}
